refactor: rollup config 공통화(모노레포간 호이스팅시킴)
