+++ .npmrc
@@ -2,3 +2,6 @@ auto-install-peers = true
 legacy-peer-deps=true
 public-hoist-pattern[]=*jest*
 public-hoist-pattern[]=*testing-library*
+public-hoist-pattern[]=*babel*
+public-hoist-pattern[]=*rollup*
+public-hoist-pattern[]=postcss-import
