+++ config/rollup-config/index.js
@@ -1,0 +1,1 @@
+export { default as rollupConfig } from './rollup.config.mjs'
