+++ packages/shared/package.json
@@ -37,18 +37,6 @@
     "convert": "node ./src/icons/profiles/generator.mjs"
   },
   "devDependencies": {
-    "@babel/core": "^7.23.9",
-    "@babel/plugin-transform-runtime": "^7.24.3",
-    "@babel/preset-env": "^7.23.9",
-    "@babel/preset-react": "^7.23.3",
-    "@babel/preset-typescript": "^7.23.3",
-    "@rollup/plugin-alias": "^5.1.0",
-    "@rollup/plugin-babel": "^6.0.4",
-    "@rollup/plugin-commonjs": "^25.0.7",
-    "@rollup/plugin-json": "^6.1.0",
-    "@rollup/plugin-node-resolve": "^15.2.3",
-    "@rollup/plugin-terser": "^0.4.4",
-    "@rollup/plugin-url": "^8.0.2",
     "@svgr/plugin-jsx": "^8.0.1",
     "@testing-library/jest-dom": "^5.16.5",
     "@testing-library/react": "^14.0.0",
@@ -63,14 +51,8 @@
     "jest-config": "workspace:*",
     "node-fetch": "^3.3.2",
     "postcss-import": "^16.1.0",
-    "rollup": "^4.14.1",
-    "rollup-plugin-import-css": "^3.5.0",
-    "rollup-plugin-node-externals": "^7.1.1",
-    "rollup-plugin-peer-deps-external": "^2.2.4",
-    "rollup-plugin-postcss": "^4.0.2",
-    "rollup-plugin-typescript2": "^0.36.0",
-    "rollup-plugin-visualizer": "^5.12.0",
     "tsconfig": "workspace:*",
+    "rollup-config": "workspace:*",
     "tslib": "^2.6.2",
     "tsup": "^5.10.1",
     "typescript": "^4.9.5"
